Feat/auth

+++ controllers/classes.py
@@ -5,7 +5,7 @@ from fastapi import Depends, Request, Response
 from fastapi.responses import HTMLResponse, RedirectResponse
 
 from dependencies import requires_owner, requires_user
-from structs import ClassRow, StudentCreate, StudentRow, UserRow
+from structs.entities import ClassRow, StudentCreate, StudentRow, UserRow
 from templates import templates
 
 
@@ -14,6 +14,9 @@ async def new(
     user: Annotated[UserRow, Depends(requires_user)]
     ) -> HTMLResponse:
     if not user:
+        if request.headers.get("Hx-Request"):
+            return Response(status_code=401, headers={"Hx-Redirect": "/"})
+        
         return RedirectResponse(status_code=303, url="/")
     
     return templates.TemplateResponse(
@@ -66,6 +69,9 @@ async def show(
     user: Annotated[UserRow, Depends(requires_owner)]
     ) -> HTMLResponse:
     if not user:
+        if request.headers.get("Hx-Request"):
+            return Response(status_code=401, headers={"Hx-Redirect": "/"})
+        
         return RedirectResponse(status_code=303, url="/")
 
     with sqlite3.connect("db.sqlite3") as conn:
@@ -80,7 +86,17 @@ async def show(
         context={"students": students}
     )
 
-async def edit(request: Request, class_id: int) -> HTMLResponse:
+async def edit(
+    request: Request,
+    class_id: int,
+    user: Annotated[UserRow, Depends(requires_owner)]
+    ) -> HTMLResponse:
+    if not user:
+        if request.headers.get("Hx-Request"):
+            return Response(status_code=401, headers={"Hx-Redirect": "/"})
+        
+        return RedirectResponse(status_code=303, url="/")
+    
     with sqlite3.connect("db.sqlite3") as conn:
         cursor = conn.cursor()
         cursor.execute(f"SELECT id, name FROM classes WHERE id = {class_id};")
@@ -95,7 +111,17 @@ async def edit(request: Request, class_id: int) -> HTMLResponse:
         context={"class": class_row, "students": students}
     )
 
-async def update(request: Request, class_id: int) -> HTMLResponse:
+async def update(
+    request: Request,
+    class_id: int,
+    user: Annotated[UserRow, Depends(requires_owner)]
+    ) -> HTMLResponse:
+    if not user:
+        if request.headers.get("Hx-Request"):
+            return Response(status_code=401, headers={"Hx-Redirect": "/"})
+        
+        return RedirectResponse(status_code=303, url="/")
+    
     form_data = await request.form()
     name = form_data.get('name')
 
@@ -120,7 +146,17 @@ async def update(request: Request, class_id: int) -> HTMLResponse:
 
     return Response(status_code=200, headers={"Hx-Refresh": "true"})
 
-async def delete(request: Request, class_id: int) -> Response:
+async def delete(
+    request: Request,
+    class_id: int,
+    user: Annotated[UserRow, Depends(requires_owner)]
+    ) -> Response:
+    if not user:
+        if request.headers.get("Hx-Request"):
+            return Response(status_code=401, headers={"Hx-Redirect": "/"})
+        
+        return RedirectResponse(status_code=303, url="/")
+    
     with sqlite3.connect("db.sqlite3") as conn:
         cursor = conn.cursor()
         cursor.execute(f"DELETE FROM classes WHERE id = ?", (class_id,))
@@ -130,8 +166,18 @@ async def delete(request: Request, class_id: int) -> Response:
     
     return RedirectResponse(status_code=303, url="/")
 
-async def students(request: Request, class_id: int) -> Response:
+async def students(
+    request: Request,
+    class_id: int,
+    user: Annotated[UserRow, Depends(requires_owner)]
+    ) -> Response:
     """Return the batch add student page"""
+    if not user:
+        if request.headers.get("Hx-Request"):
+            return Response(status_code=401, headers={"Hx-Redirect": "/"})
+        
+        return RedirectResponse(status_code=303, url="/")
+
     with sqlite3.connect("db.sqlite3") as conn:
         cursor = conn.cursor()
         cursor.execute(f"SELECT id, name FROM classes WHERE id = {class_id};")
@@ -144,7 +190,17 @@ async def students(request: Request, class_id: int) -> Response:
     )
     
 
-async def create_batch(request: Request, class_id: int) -> HTMLResponse:
+async def create_batch(
+    request: Request,
+    class_id: int,
+    user: Annotated[UserRow, Depends(requires_owner)]
+    ) -> HTMLResponse:
+    if not user:
+        if request.headers.get("Hx-Request"):
+            return Response(status_code=401, headers={"Hx-Redirect": "/"})
+        
+        return RedirectResponse(status_code=303, url="/")
+    
     form_data = await request.form()
     students = form_data.get("students")
     
@@ -166,7 +222,19 @@ async def create_batch(request: Request, class_id: int) -> HTMLResponse:
     return RedirectResponse(status_code=303, url=f"/classes/{class_id}/edit")
 
 
-async def edit_student(request: Request, class_id: int, student_id: int):
+async def edit_student(
+    request: Request,
+    class_id: int,
+    student_id: int,
+    user: Annotated[UserRow, Depends(requires_owner)]
+    ) -> HTMLResponse:
+    """Returns the inline edit student form"""
+    if not user:
+        if request.headers.get("Hx-Request"):
+            return Response(status_code=401, headers={"Hx-Redirect": "/"})
+        
+        return RedirectResponse(status_code=303, url="/")
+    
     with sqlite3.connect("db.sqlite3") as conn:
         cursor = conn.cursor()
         
@@ -181,3 +249,50 @@ async def edit_student(request: Request, class_id: int, student_id: int):
         name="classes/_edit-student.html",
         context={"class": class_row, "student": student_row}
     )
+
+
+async def update_student(
+    request: Request,
+    student_id: int,
+    user: Annotated[UserRow, Depends(requires_owner)]
+    ) -> Response:
+    """
+    Updates a student resource.
+    Only exposes updating the name as students can't be moved between classes.
+    Triggers a page refresh with the response.
+    """
+    if not user:
+        if request.headers.get("Hx-Request"):
+            return Response(status_code=401, headers={"Hx-Redirect": "/"})
+        
+        return RedirectResponse(status_code=303, url="/")
+    
+    form_data = await request.form()
+    name = form_data.get("student")
+
+    with sqlite3.connect("db.sqlite3") as conn:
+        cursor = conn.cursor()
+        cursor.execute("UPDATE STUDENTS SET name = ? WHERE id = ?", (name, student_id,))
+
+    return Response(status_code=200, headers={"Hx-Refresh": "true"})
+
+
+async def delete_student(
+    request: Request,
+    student_id: int,
+    user: Annotated[UserRow, Depends(requires_owner)]
+    ) -> Response:
+    """
+    Deletes a student resource.
+    Responds with a response. No visible client side confirmation except the student disappearing, for now.
+    """
+    if not user:
+        if request.headers.get("Hx-Request"):
+            return Response(status_code=401, headers={"Hx-Redirect": "/"})
+        
+        return RedirectResponse(status_code=303, url="/")
+    with sqlite3.connect("db.sqlite3") as conn:
+        cursor = conn.cursor()
+        cursor.execute("DELETE FROM students WHERE id = ?", (student_id,))
+
+    return Response(status_code=200, content="Student deleted successfully")
